+++ core/src/server/delivery/hybrid_request/CreateHybridCollectionRequest.cpp
@@ -75,6 +75,11 @@ CreateHybridCollectionRequest::OnExecute() {
         for (auto& field_type : field_types_) {
             engine::meta::hybrid::FieldSchema schema;
             auto field_name = field_type.first;
+            status = ValidateFieldName(field_name);
+            if (!status.ok()) {
+                return status;
+            }
+
             auto index_params = field_index_params_.at(field_name);
             schema.collection_id_ = collection_name_;
             schema.field_name_ = field_name;
@@ -103,7 +108,12 @@ CreateHybridCollectionRequest::OnExecute() {
         collection_info.collection_id_ = collection_name_;
         collection_info.dimension_ = dimension;
         if (extra_params_.contains("segment_size")) {
-            collection_info.index_file_size_ = extra_params_["segment_size"].get<int64_t>();
+            auto segment_size = extra_params_["segment_size"].get<int64_t>();
+            collection_info.index_file_size_ = segment_size;
+            status = ValidateCollectionIndexFileSize(segment_size);
+            if (!status.ok()) {
+                return status;
+            }
         }
 
         if (vector_param.contains("metric_type")) {