+++ CHANGELOG.md
@@ -16,6 +16,10 @@ Please mark all changes in change log and use the issue from GitHub
 -   \#2637 Suit the range of HNSW parameters
 -   \#2649 Search parameter of annoy has conflict with document
 -   \#2693 Collection create success if no dimension value provided
+-   \#2694 Collection create success if an invalid field name provided
+-   \#2695 The number of fields should be limited
+-   \#2696 Check the validity of the parameters of creating collection: segment_size
+-   \#2697 Index can not be created
 
 ## Feature
 -   \#2319 Redo metadata to support MVCC